requirements.txt
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
requests==2.27.1
